+++ src/phys/time/mod.rs
@@ -8,10 +8,12 @@
 #![cfg_attr(feature = "safe_time", forbid(unsafe_code))]
 
 mod delta; // TimeDelta
+// mod drop; // TimeDrop
 mod fmt; // Timecode
-mod granularity; // TimeGranularity
+// mod freq; // TimeFreq
 mod no; // NoTime
 mod reexports; // std::time::*
+mod scale; // TimeScale
 mod source; // TimeSource
 mod split; // TimeSplit[Year[Day|Sec]|Hour[Sec|Nano]|MilliNano][Norm]
 
@@ -21,9 +23,6 @@ crate::items! {
     mod unix; // TimeUnix[I64|U32]
 }
 
-// WIPZONE
-// mod drop;
-// mod freq;
 // #[cfg(feature = "_destaque_u16")]
 // mod looper;
 // mod instant;
@@ -32,10 +31,12 @@ crate::structural_mods! { // _mods
     _mods {
         pub use super::{
             delta::*,
+            // drop::*,
             fmt::*,
-            granularity::*,
+            // freq::*;
             no::*,
             reexports::*,
+            scale::*,
             source::*,
             split::*,
         };
@@ -47,8 +48,6 @@ crate::structural_mods! { // _mods
             unix::*,
         };
         // WIPZONE
-        // pub use super::drop::*;
-        // pub use super::freq::*;
         // #[cfg(feature = "_destaque_u16")]
         // #[cfg_attr(nightly_doc, doc(cfg(feature = "_destaque_u16")))]
         // pub use super::looper::*;